Revert "wip"
This reverts commit c2b6eda8778f4f241668880433c74d70f47d46f5.

+++ SpringBootDeneme/src/main/java/com/example/SpringBootDeneme/Repository/PatientsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface PatientsRepository extends JpaRepository<Patients, Long> {
-    List<Patients> findByPatientId(long patientId);
+    List<Patients> findByPatientId(Integer patientId);
 }
